Handle extended CCs 140 and 141 (keydelta)

+++ src/sfizz/Synth.cpp
@@ -348,7 +348,7 @@ void Synth::Impl::handleGlobalOpcodes(const std::vector<Opcode>& members)
 
 void Synth::Impl::handleGroupOpcodes(const std::vector<Opcode>& members, const std::vector<Opcode>& masterMembers)
 {
-    absl::optional<int> groupIdx;
+    absl::optional<int64_t> groupIdx;
     absl::optional<unsigned> maxPolyphony;
 
     const auto parseOpcode = [&](const Opcode& rawMember) {
